Make rootrw upperdir customizable
This is to make it possible for user to define their own upperdir location
in rootrw.
This gives the user more control over the upperdir(s) and the possiblity
to manage multiple versions of upperdir.

One use case behind this is:
E.g., if the user has A/B partition scheme, A & B RO rootfs,
but only have one disk/volume for the rootrw due to limited partitioning space,
and the user wants to switch between A/B RO rootfs meanwhile keep the overlayfs
matching the RO rootfs that is being activated,
then the user can define their custom upperdir name via kernel cmdline.